Fixing output and refactor more
